# backend/swagger.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/api/faculty/applications/{id}/stage": {
            "put": {
                "tags": [
                    "faculty-application-controller"
                ],
                "operationId": "updateApplicationStage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "facultyEmail",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "faculty@dept.com"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StageUpdateRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/faculty/events": {
            "get": {
                "tags": [
                    "faculty-announcement-controller"
                ],
                "operationId": "getEvents",
                "parameters": [
                    {
                        "name": "facultyEmail",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "faculty@dept.com"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseListEvent"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "faculty-announcement-controller"
                ],
                "operationId": "createEvent",
                "parameters": [
                    {
                        "name": "facultyEmail",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "faculty@dept.com"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Event"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseEvent"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/faculty/announcements": {
            "get": {
                "tags": [
                    "faculty-announcement-controller"
                ],
                "operationId": "getAnnouncements",
                "parameters": [
                    {
                        "name": "facultyEmail",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "faculty@dept.com"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseListAnnouncement"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "faculty-announcement-controller"
                ],
                "operationId": "createAnnouncement",
                "parameters": [
                    {
                        "name": "facultyEmail",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "faculty@dept.com"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Announcement"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseAnnouncement"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/admin/announcements": {
            "get": {
                "tags": [
                    "admin-announcement-controller"
                ],
                "operationId": "getAllAnnouncements",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseListAnnouncementResponseDTO"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "admin-announcement-controller"
                ],
                "operationId": "createAnnouncement_1",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "admin@campushire.com"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AnnouncementRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseAnnouncementResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/faculty/applications": {
            "get": {
                "tags": [
                    "faculty-application-controller"
                ],
                "operationId": "getDepartmentApplications",
                "parameters": [
                    {
                        "name": "facultyEmail",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "faculty@dept.com"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseListFacultyApplicationDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/faculty/events/{id}": {
            "delete": {
                "tags": [
                    "faculty-announcement-controller"
                ],
                "operationId": "deleteEvent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "facultyEmail",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "faculty@dept.com"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/faculty/announcements/{id}": {
            "delete": {
                "tags": [
                    "faculty-announcement-controller"
                ],
                "operationId": "deleteAnnouncement",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "facultyEmail",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "faculty@dept.com"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/announcements/{id}": {
            "delete": {
                "tags": [
                    "admin-announcement-controller"
                ],
                "operationId": "deleteAnnouncement_1",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseString"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "ApiResponseVoid": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "object"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "StageUpdateRequestDTO": {
                "type": "object",
                "properties": {
                    "targetStage": {
                        "type": "string",
                        "enum": [
                            "APPLIED",
                            "ASSESSMENT",
                            "TECHNICAL",
                            "HR",
                            "SELECTED",
                            "REJECTED"
                        ]
                    }
                },
                "required": [
                    "targetStage"
                ]
            },
...
            "Department": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "code": {
                        "type": "string"
                    }
                }
            },
            "Event": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "scheduledAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "locationOrLink": {
                        "type": "string"
                    },
                    "department": {
                        "$ref": "#/components/schemas/Department"
                    }
                }
            },
            "ApiResponseEvent": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/Event"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "Announcement": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    },
                    "scope": {
                        "type": "string",
                        "enum": [
                            "GLOBAL",
                            "DEPARTMENT"
                        ]
                    },
                    "department": {
                        "$ref": "#/components/schemas/Department"
                    },
                    "createdBy": {
                        "$ref": "#/components/schemas/User"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "email": {
                        "type": "string"
                    },
                    "universityRegNo": {
                        "type": "string"
                    },
                    "passwordHash": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "STUDENT",
                            "FACULTY",
                            "PLACEMENT_HEAD"
                        ]
                    },
                    "department": {
                        "$ref": "#/components/schemas/Department"
                    },
                    "isActive": {
                        "type": "boolean"
                    }
                }
            },
            "ApiResponseAnnouncement": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/Announcement"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "AnnouncementRequestDTO": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    },
                    "scope": {
                        "type": "string",
                        "enum": [
                            "GLOBAL",
                            "DEPARTMENT"
                        ]
                    },
                    "departmentId": {
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "required": [
                    "content",
                    "scope",
                    "title"
                ]
            },
            "AnnouncementResponseDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    },
                    "scope": {
                        "type": "string",
                        "enum": [
                            "GLOBAL",
                            "DEPARTMENT"
                        ]
                    },
                    "departmentName": {
                        "type": "string"
                    },
                    "createdByEmail": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "ApiResponseAnnouncementResponseDTO": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/AnnouncementResponseDTO"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "ApiResponseString": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "string"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "ApiResponseListEvent": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Event"
                        }
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "ApiResponseListAnnouncement": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Announcement"
                        }
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "ApiResponseListFacultyApplicationDTO": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/FacultyApplicationDTO"
                        }
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "FacultyApplicationDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "studentId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "studentName": {
                        "type": "string"
                    },
                    "rollNo": {
                        "type": "string"
                    },
                    "driveId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "companyName": {
                        "type": "string"
                    },
                    "driveRole": {
                        "type": "string"
                    },
                    "stage": {
                        "type": "string",
                        "enum": [
                            "APPLIED",
                            "ASSESSMENT",
                            "TECHNICAL",
                            "HR",
                            "SELECTED",
                            "REJECTED"
                        ]
                    },
                    "appliedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "lastUpdatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "ApiResponseListAnnouncementResponseDTO": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AnnouncementResponseDTO"
                        }
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
        }
    }
}